Authorization update  + Gast klasse + Gast access

+++ src/main/java/be/kdg/teamh/controllers/CirkelsessieController.java
@@ -3,7 +3,9 @@ package be.kdg.teamh.controllers;
 import be.kdg.teamh.entities.*;
 import be.kdg.teamh.exceptions.CirkelsessieNotFound;
 import be.kdg.teamh.exceptions.GebruikerNotFound;
+import be.kdg.teamh.exceptions.IsForbidden;
 import be.kdg.teamh.exceptions.SubthemaNotFound;
+import be.kdg.teamh.services.contracts.AuthService;
 import be.kdg.teamh.services.contracts.CirkelsessieService;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
@@ -19,21 +21,33 @@ public class CirkelsessieController {
     @Autowired
     private CirkelsessieService service;
 
+    @Autowired
+    private AuthService auth;
+
     @ResponseStatus(HttpStatus.OK)
     @RequestMapping(value = "", method = RequestMethod.GET)
-    public List<Cirkelsessie> index() {
+    public List<Cirkelsessie> index() throws IsForbidden {
+//        if (!auth.isRegistered(token))
+//            throw new IsForbidden();
         return service.all();
     }
 
     @ResponseStatus(HttpStatus.CREATED)
     @RequestMapping(value = "", method = RequestMethod.POST)
-    public void create(@RequestBody Cirkelsessie cirkelsessie, @RequestHeader(name = "Authorization") String token) {
+    public void create(@RequestBody Cirkelsessie cirkelsessie,
+                       @RequestHeader(name = "Authorization") String token) throws IsForbidden {
+        if (!auth.isRegistered(token))
+            throw new IsForbidden();
         service.create(getUserId(token), cirkelsessie);
     }
 
     @ResponseStatus(code = HttpStatus.CREATED)
     @RequestMapping(value = "subthema={id}", method = RequestMethod.POST)
-    public void create(@RequestBody Cirkelsessie cirkelsessie, @PathVariable("id") int subthemaId, @RequestHeader(name = "Authorization") String token) throws GebruikerNotFound, SubthemaNotFound {
+    public void create(@RequestBody Cirkelsessie cirkelsessie, @PathVariable("id") int subthemaId,
+                       @RequestHeader(name = "Authorization") String token)
+            throws GebruikerNotFound, SubthemaNotFound, IsForbidden {
+        if (!auth.isRegistered(token))
+            throw new IsForbidden();
         service.create(getUserId(token), subthemaId, cirkelsessie);
     }
 
@@ -46,13 +60,28 @@ public class CirkelsessieController {
 
     @ResponseStatus(HttpStatus.OK)
     @RequestMapping(value = "{id}", method = RequestMethod.PUT)
-    public void update(@PathVariable("id") int id, @RequestBody Cirkelsessie cirkelsessie) throws CirkelsessieNotFound {
+    public void update(@PathVariable("id") int id, @RequestBody Cirkelsessie cirkelsessie,
+                       @RequestHeader(name = "Authorization") String token,
+                       @RequestParam(name = "t", required = false) String tokenParam)
+            throws CirkelsessieNotFound, IsForbidden, GebruikerNotFound {
+        if (tokenParam != null && auth.isGuest(tokenParam)) {
+            Gast gast = auth.findGastByToken(tokenParam);
+            List<Cirkelsessie> cirkelsessies = gast.getCirkelsessies();
+            if (cirkelsessies.contains(cirkelsessie))
+                service.update(id, cirkelsessie);
+            else throw new IsForbidden();
+        }
+        else if (!auth.isRegistered(token) && !auth.isGuest(tokenParam))
+            throw new IsForbidden();
         service.update(id, cirkelsessie);
     }
 
     @ResponseStatus(HttpStatus.OK)
     @RequestMapping(value = "{id}", method = RequestMethod.DELETE)
-    public void delete(@PathVariable("id") int id) throws CirkelsessieNotFound {
+    public void delete(@PathVariable("id") int id, @RequestHeader(name = "Authorization") String token)
+            throws CirkelsessieNotFound, IsForbidden {
+        if (!auth.isRegistered(token))
+            throw new IsForbidden();
         service.delete(id);
     }
 
@@ -64,7 +93,10 @@ public class CirkelsessieController {
 
     @ResponseStatus(HttpStatus.CREATED)
     @RequestMapping(value = "{id}/clone", method = RequestMethod.POST)
-    public void clone(@PathVariable("id") int id) throws CirkelsessieNotFound {
+    public void clone(@PathVariable("id") int id, @RequestHeader(name = "Authorization") String token)
+            throws CirkelsessieNotFound, IsForbidden {
+        if (!auth.isRegistered(token))
+            throw new IsForbidden();
         service.clone(id);
     }
 
@@ -82,8 +114,9 @@ public class CirkelsessieController {
 
     @ResponseStatus(HttpStatus.CREATED)
     @RequestMapping(value = "{id}/spelkaart", method = RequestMethod.POST)
-    public void createKaart(@PathVariable("id") int id, @RequestHeader(name = "Authorization") String token, @RequestBody Kaart kaart) throws CirkelsessieNotFound, GebruikerNotFound {
-
+    public void createKaart(@PathVariable("id") int id, @RequestHeader(name = "Authorization") String token, @RequestBody Kaart kaart) throws CirkelsessieNotFound, GebruikerNotFound, IsForbidden {
+        if (!auth.isRegistered(token))
+            throw new IsForbidden();
         service.addSpelkaart(id, getUserId(token), kaart);
     }
